Separate yac mock and test
